# Fix error checking in lyapunov and lyapunovspectrum