This is an attempt to fix issue [184](https://github.com/JuliaDynamics/DynamicalSystems.jl/issues/184). I've added error checking: `lyapunov` and `lyapunovspectrum` now return NaN values immediately in the case of instability. Note that this version only tests for the `:Unstable` retcode and only if the integrator is a usual one. In case of the simple integrator types the `dt<dtmin` error is raised on instability. 
I also did some benchmarking on `Systems.lorenz`.
With error checking:
```
@btime lyapunov(ds,100, Ttr = 2000 ;diffeq)
  23.199 ms (666 allocations: 50.45 KiB)
0.9056821285687465

@btime lyapunovspectrum(ds,100, Ttr = 2000 ;diffeq)
  27.496 ms (57 allocations: 6.91 KiB)
3-element Vector{Float64}:
   0.9147419224955904
  -0.005725267193229689
 -14.575596505931658

```

Without error checking:
```
 @btime lyapunov(ds,100, Ttr = 2000 ;diffeq)
  23.201 ms (666 allocations: 50.45 KiB)
0.9056821285687465

@btime lyapunovspectrum(ds,100, Ttr = 2000 ;diffeq)
  27.469 ms (57 allocations: 6.91 KiB)
3-element Vector{Float64}:
   0.9147419224955904
  -0.005725267193229689
 -14.575596505931658


```