@Datseris 2022-07-03
Thank you very much!

Can you please open a PR at DynamicalSystemsBase.jl, in which you define a function
```
successful_step(integ)
```
which returns `true` if the latest step the integrator took was successful? For integrators that don't have this return code field the function will always return true. Then you can use that function in this code here, which will lead to cleaner code as there would be no reason to initialize `has_retcode` etc.

Furthermore, instead of `length(get_state(ds))` use `dimension(ds)`.

lastly, you need to add a test in the test suite that indeed confirms that we get `NaN`s for unstble initial conditions.