@rusandris 2022-07-13
Hi! I've modified this pull request. Now it's much cleaner with the suggested `successful_step` functions. And also, I hope it works for all simple integrators and tests for all retcodes, not just `:Unstable`. I've tested the changes locally, but I need help with this

> lastly, you need to add a test in the test suite that indeed confirms that we get NaNs for unstble initial conditions.