@Datseris 2022-07-13
Thanks, I'll go through the other PR now and give some help.

For the tests of the current PR, you need to add a test that computes the lyapunovspectrum for the unstable system-parameters mentioned in the issue and ensure that the return is NaN.