@Jaderpolli 2024-06-07
Hi! I am facing this same problem, but the implemented solution on this post is not on the current version of lyapunov.jl.
If you try this minimal code here

```
using DynamicalSystems, DifferentialEquations

function fHH!(du, u, p, t)
    du[1] = (p[4] .+ p[2] * sin.(2*pi*p[1]*t) - p[8]*u[2]^(3)*u[3]*(u[1] - p[5]) - p[9]*u[4]^4*(u[1] - p[6]) - p[10]*(u[1] - p[7]))/p[3]
    du[2] = 0.1 * (u[1] + 40.0)/(1.0 - exp(-(u[1] + 40.0) / 10.0))* (1.0 - u[2]) - 4.0 * exp(-(u[1] + 65.0) / 18.0) * u[2]
    du[3] = 0.07 * exp(-(u[1] + 65.0) / 20.0) * (1.0 - u[3]) - 1.0 / (1 + exp(-(u[1] + 35.0) / 10.0)) * u[3]
    du[4] = 0.01 * (u[1] + 55.0) / (1.0 - exp(-(u[1] + 55.0) / 10.0)) * (1.0 - u[4]) - 0.125 * exp(-(u[1] + 65.0) / 80.0) * u[4]
    return nothing
end

function testlyap()
    Cm  =   1.0 # membrane capacitance, in uF/cm^2
    g_Na = 120.0 # maximum conducances, in mS/cm^2
    g_K  =  36.0
    g_L  =   0.3
    E_Na =  50.0 # Nernst reversal potentials, in mV
    E_K  = -77.0
    E_L  = -54.387
    I0 = 0.0
    Astim = 6660
    fStim = 4.0 
    u0 = [-6.49963792e+01, 5.29550879e-02, 5.95994171e-01, 3.17732402e-01]
    diffeq =  (alg = Tsit5(), abstol = 1e-7, reltol = 1e-7)
    p0 = [fStim, Astim, Cm, I0, E_Na, E_K, E_L, g_Na, g_K, g_L]
    forcedHH = ContinuousDynamicalSystem(fHH!, u0, p0; diffeq)      
    return lyapunov(forcedHH, 10; Ttr = 0, Δt = 0.5)
end

testlyap()
```

With this specific `abstol` and `reltol` it returns a warning and simple takes too long to finish. In fact on the example here for T = 10 ms is almost instantaneous, but it returns the error `After rescaling, the distance of reference and test states was not d0_lower ≤ d ≤ d0_upper as expected. Perhaps you are using a dynamical system where the algorithm doesn't work` which occurs because `d` becomes zero and the program shuts down. 
This is a secondary problem that might also just return a `NaN` since we have coalescent orbits.

When I update my local file `lyapunov.jl` with lines for testing `d=0` and the `successful_step(pds) || return NaN` everything works fine. Of course, for some cases the problems just occur for specific parameter, integrator (I tried a lot with ``Rosenbrock23` to account stiffness but still have same problem) and `abstol`, `reltol` values, but since I'm running over a large variety of parameters, I need to set these beforehand.

So, the question is: is this not implemented because it is provisional and we must resolve [184](https://github.com/JuliaDynamics/DynamicalSystems.jl/issues/184) or any other reason? Is there a problem with changing only my local files, thinking about reproducibility?