@Datseris 2024-06-07
Yes, all functions should have the `successful_step(ds) || return NaN` option when sensible. `lyapunov` is one. I think I closed this because it was opened in DynamicalSystems.jl v2 and now we are in v3, so the PR would need a rework. Feel free to open a new one that does the same thing. Add your example system above as a test case in the test folder. And that should be it!